requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
anthropic>=0.39.0
apscheduler>=3.10.0
pandas>=2.1.0
python-dotenv>=1.0.0
